# Perf: use Queue.length instead Seq.length