before

| Method |    Mean |   Error |  StdDev | Rank |        Gen 0 |       Gen 1 |       Gen 2 | Allocated |
|------- |--------:|--------:|--------:|-----:|-------------:|------------:|------------:|----------:|
| Format | 41.41 s | 21.28 s | 1.166 s |    1 | 2660000.0000 | 920000.0000 | 148000.0000 |   13.1 GB |

after

| Method |    Mean |   Error |  StdDev | Rank |        Gen 0 |       Gen 1 |      Gen 2 | Allocated |
|------- |--------:|--------:|--------:|-----:|-------------:|------------:|-----------:|----------:|
| Format | 37.33 s | 2.964 s | 0.162 s |    1 | 2357000.0000 | 759000.0000 | 66000.0000 |  11.92 GB |